Further implementations

+++ Afrodite/Abstract/IComponentProperties.cs
@@ -11,12 +11,12 @@
 		/// Gets or sets the cpu cores count.
 		/// </summary>
 		/// <value>The cpu cores count.</value>
-		int CpuCoresCount{get; set;}
+		uint CpuCoresCount{get; set;}
 		/// <summary>
 		/// Gets or sets the total machine RAM memory in MB.
 		/// </summary>
 		/// <value>The total machine RAM memory in MB.</value>
-		int TotalMemoryMBytes { get; set;}
+        ulong TotalMemoryMBytes { get; set; }
 		/// <summary>
 		/// Gets or sets the info about OS.
 		/// </summary>
@@ -26,7 +26,7 @@
 		/// Gets or sets the total disk capcity in MB .
 		/// </summary>
 		/// <value>The total disk capcity in MB.</value>
-		int TotalDiskMBytes {get;set;}
+		ulong TotalDiskMBytes {get;set;}
 		/// <summary>
 		/// Gets or sets the name of the coputer.
 		/// </summary>
